Removed unused vue logo, moved demo data into separate demo file
