+++ src/demo.js
@@ -1,0 +1,58 @@
+import moment from 'moment'
+
+export const servers = [{
+  name: 'www01',
+  status: 'success',
+  icon: 'globe',
+  description: 'Web server that runs our sites'
+}, {
+  name: 'sql01',
+  status: 'danger',
+  icon: 'database',
+  description: 'mySQL server used for reporting'
+}, {
+  name: 'mongoDB01',
+  status: 'info',
+  icon: 'file-code-o',
+  description: 'Main DB server'
+}, {
+  name: 'ldap01',
+  status: 'success',
+  icon: 'key',
+  description: 'Authentication server'
+}, {
+  name: 'mgmt01',
+  status: 'success',
+  icon: 'home',
+  description: 'Management server with all tools'
+}, {
+  name: 'bkup01',
+  status: 'warning',
+  icon: 'backward',
+  description: 'Backup server'
+}]
+
+export const timeline = [{
+  icon: 'fa-envelope',
+  color: 'blue',
+  title: 'Write short novel',
+  time: moment().endOf('day').fromNow(),
+  body: 'Etsy doostang zoodles disqus groupon greplin oooj voxy zoodles, weebly ning heekya handango imeem plugg dopplr jibjab, movity jajah plickers sifteo edmodo ifttt zimbra. Babblely odeo kaboodle quora plaxo ideeli hulu weebly balihoo...',
+  buttons: [{
+    type: 'primary',
+    message: 'Read more'
+  }]
+},
+{
+  icon: 'fa-user',
+  color: 'yellow',
+  title: 'Sarah Young accepted your friend request',
+  time: moment('20150620', 'MMM Do YY').fromNow()
+},
+{
+  icon: 'fa-camera',
+  color: 'purple',
+  title: 'Watch a youTube video',
+  time: moment('20130620', 'YYYYMMDD').fromNow(),
+  body: '<div class="embed-responsive embed-responsive-16by9"><iframe width="560" height="315" src="https://www.youtube.com/embed/8aGhZQkoFbQ" frameborder="0" allowfullscreen></iframe></div>'
+}]
